move a couple tests from core to parse
